pamparam/requirements.txt
python-binance
ta
json
numpy
pandas
Flask